+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 tweakStore 0.x.x
-- create themes tab
 - create update from external repositories
 - fix permissions
 - fix translation
+
+tweakStores 0.0.x - 2021.08.xx
+- add write dcstore.xml for modules tab
+- create themes tab
 
 tweakStores 0.0.2 - 2021.08.26
 - initial commit